requirements.txt
numpy~=1.23.1
PyAutoGUI~=0.9.53
opencv-python~=4.6.0.66
Pillow~=9.2.0
easyocr~=1.5.0